# 🚀 Boot code optimization: store as base64 string