**Boot code size**:

| File Name           | Before | After  | Change |
|---------------------|--------|--------|----------|
| bootcode.js         | 8.5 MB | 3.3 MB | 61% ▼    |
| bootcode.browser.js | 8.8 MB | 3.2 MB | 63% ▼    |


**Boot time**:

|       | Before     | After      | % Change |
|-------|------------|------------|----------|
| load  | 344.4 ms   | 171.458 ms | 50% ▼    |
| boot  | 502.711 ms | 206.429 ms | 58% ▼    |
| ready | 848.932 ms | 379.64 ms  | 55%  ▼    |

<details>
  <summary>Source</summary>

```js
console.time('ready');
console.time('load');
var Sandbox = require('postman-sandbox');
console.timeEnd('load');

console.time('boot');
Sandbox.createContext(function (err, ctx) {
  console.timeEnd('boot');
  console.timeEnd('ready');
});
```
</details>

**Tokens**: (`esprima.tokenize`)


| | Before | After  |
|-|-------|-------|
|Tokens| 5155340 | 25 |